requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Security-Too==5.3.2
Flask-WTF==1.2.1
GeoAlchemy2==0.14.3
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.4
packaging==23.2
passlib==1.7.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
shapely==2.0.2
SQLAlchemy==2.0.23
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.1
zipp==3.17.0
